Pulled `Tiles` out of `MainActivity`
Preview where?
